requirements.txt
aiohttp==3.7.4.post0
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
backcall==0.2.0
certifi==2020.12.5
chardet==4.0.0
decorator==4.4.2
Django==3.1.7
dnspython==2.0.0
flake8==3.9.0
geoip2==4.1.0
idna==2.10
ipwhois==1.2.0
IPy==1.1
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
maxminddb==2.0.3
mccabe==0.6.1
multidict==5.1.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.0
pygeoip==0.3.2
Pygments==2.8.1
pytz==2021.1
requests==2.25.1
scapy==2.4.4
sqlparse==0.4.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
yarl==1.6.3
